+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 Note: This is in reverse chronological order, so newer entries are added to the top.
 
+Swift Next
+-----------
+* [#5949]
+  
+  New `--pkg-config-path` option on `build`, `test`, and `run` commands has been
+  introduced as an alternative to passing `PKG_CONFIG_PATH` environment variable.
+  It allows specifying alternative path to search for `.pc` files used by
+  `pkg-config`. Use the option multiple times to specify more than one path.
+
 Swift 5.8
 -----------